Convert from packer to lazy
