--- ftdetect/cmake.vim
@@ -1,2 +1,0 @@
-autocmd BufRead,BufNewFile CMakeLists.txt set filetype=cmake
-autocmd BufEnter *.cpp,*.h,*.hpp :lua vim.api.nvim_buf_set_option(0, "commentstring", "// %s")
