--- ftdetect/lua.vim
@@ -1,1 +1,0 @@
-autocmd BufEnter *.lua :setlocal shiftwidth=2 tabstop=2
