+++ init.lua
@@ -1,15 +1,28 @@
-require('lego.options')
-require('lego.plugins')
-require('lego.mason-config')
-require('lego.lsp-clients')
-require('lego.treesitter')
-require('lego.comment')
-require('lego.keymaps')
-require('lego.telescope') -- must be after keymaps
-require('lego.lualine')
-require('lego.gitsigns')
-require('lego.which-key')
-require('lego.harpoon')
+vim.g.mapleader = " "
+vim.g.maplocalleader = " "
 
-require('lego.nvimtree')
+local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
+if not (vim.uv or vim.loop).fs_stat(lazypath) then
+  vim.fn.system({
+    "git",
+    "clone",
+    "--filter=blob:none",
+    "https://github.com/folke/lazy.nvim.git",
+    "--branch=stable", -- latest stable release
+    lazypath,
+  })
+end
+vim.opt.rtp:prepend(lazypath)
+
+require "lazy" .setup {
+  "folke/which-key.nvim",
+  { "folke/neoconf.nvim", cmd = "Neoconf" },
+  "folke/neodev.nvim",
+  { import = "plugins" },
+}
+
+require "core.autocmd"
+require "core.colors"
+require "core.options"
+require "core.keymaps"
