+++ lua/core/colors.lua
@@ -1,0 +1,5 @@
+vim.opt.termguicolors = true
+vim.cmd 'colorscheme habamax'
+vim.api.nvim_command('hi WinSeparator guibg=NONE ctermbg=NONE')
+-- vim.api.nvim_command('hi Normal guibg=NONE ctermbg=NONE')
+
