+++ lua/core/options.lua
@@ -1,0 +1,9 @@
+vim.opt.scrolloff = 10
+vim.opt.tabstop = 4
+vim.opt.shiftwidth = 4
+vim.opt.expandtab = true
+vim.opt.number = true
+vim.opt.numberwidth = 5
+vim.opt.cursorline = true
+
+
