--- lua/lego/lualine.lua
@@ -1,8 +1,0 @@
-vim.o.showmode = false
-require('lualine').setup {
-  options = {
-    globalstatus = true,
-    -- section_separators = {'', ''},
-  },
-}
-
