--- lua/lego/mason-config.lua
@@ -1,4 +1,0 @@
-require("mason").setup()
-require("mason-lspconfig").setup{
-  ensure_installed = { "lua_ls", "clangd", "neocmake" },
-}
