+++ lua/plugins/debugger.lua
@@ -1,0 +1,15 @@
+return {
+  "mfussenegger/nvim-dap",
+  dependencies={
+--    "williamboman/mason.nvim",
+--    "nvim-neotest/nvim-nio",
+--    "rcarriga/nvim-dap-ui",
+--    "theHamsta/nvim-dap-virtual-text",
+  },
+  config=function ()
+--    require("dap").setup()
+--    require("dapui").setup()
+--    require("nvim-dap-virtual-text").setup { }
+  end
+}
+
