+++ lua/plugins/lualine.lua
@@ -1,0 +1,11 @@
+return {
+  "nvim-lualine/lualine.nvim",
+  config = function()
+    vim.o.showmode = false
+    require "lualine".setup {
+      options = {
+        globalstatus = true,
+      }
+    }
+  end
+}
