+++ lua/plugins/mason.lua
@@ -1,0 +1,16 @@
+return {
+    "williamboman/mason.nvim",
+    event = "VeryLazy",
+    config = function()
+        require "mason".setup{
+            ui = {
+                icons = {
+                    package_installed = "✓",
+                    package_pending = "➜",
+                    package_uninstalled = "✗"
+                }
+            }
+        }
+    end
+}
+
