+++ lua/plugins/nvim-lspconfig.lua
@@ -8,11 +8,6 @@ local lsp_flags = function(desc)
   return res
 end
 
-vim.keymap.set('n', '<space>w', vim.diagnostic.open_float, lsp_flags 'Expand disagnostic in float')
-vim.keymap.set('n', ']d', vim.diagnostic.goto_next,        lsp_flags 'Go to next diagnostic')
-vim.keymap.set('n', '[d', vim.diagnostic.goto_prev,        lsp_flags 'Go to previous diagnostic')
-vim.keymap.set('n', '<space>q', vim.diagnostic.setloclist, lsp_flags 'Set loclist')
-
 -- Use an on_attach function to only map the following keys
 -- after the language server attaches to the current buffer
 local on_attach = function(client, bufnr)
@@ -42,55 +37,66 @@ local on_attach = function(client, bufnr)
   vim.keymap.set('n', 'gr', vim.lsp.buf.references,            bufopts 'Show references')
   vim.keymap.set('n', '<space>f', function() vim.lsp.buf.format { async = true } end, bufopts 'Format file')
 
-  --vim.keymap.set('n', '<space>dn', vim.diagnostic.goto_next, bufopts)
+  vim.keymap.set('n', '<space>w', vim.diagnostic.open_float, lsp_flags 'Expand disagnostic in float')
+  vim.keymap.set('n', ']d', vim.diagnostic.goto_next,        lsp_flags 'Go to next diagnostic')
+  vim.keymap.set('n', '[d', vim.diagnostic.goto_prev,        lsp_flags 'Go to previous diagnostic')
+  vim.keymap.set('n', '<space>q', vim.diagnostic.setloclist, lsp_flags 'Set loclist')
 end
 
-require('lspconfig')['pyright'].setup{
+local lsp_config = function()
+
+  require('lspconfig')['pyright'].setup{
     on_attach = on_attach,
     flags = lsp_flags(),
     root_dir = function()
         return vim.fs.dirname(vim.fs.find({'setup.py'}, { upward = true })[1])
     end,
-}
-require('lspconfig')['tsserver'].setup{
+  }
+  require('lspconfig')['tsserver'].setup{
     on_attach = on_attach,
     flags = lsp_flags(),
-}
--- Install language server: 'cargo install neocmakelsp'
-require'lspconfig'.neocmake.setup{
+  }
+  -- Install language server: 'cargo install neocmakelsp'
+  require'lspconfig'.neocmake.setup{
     on_attach = on_attach,
     flags = lsp_flags(),
-}
--- To install language server:
---  * On ubuntu 'sudo apt install clangd'
---  * On Windows, make sure that clangd.exe from LLVM is in the path or use the one bundled with Visual Studio
-require'lspconfig'.clangd.setup{
+  }
+  -- To install language server:
+  --  * On ubuntu 'sudo apt install clangd'
+  --  * On Windows, make sure that clangd.exe from LLVM is in the path or use the one bundled with Visual Studio
+  require'lspconfig'.clangd.setup{
     on_attach = on_attach,
     flags = lsp_flags(),
-    cmd = (function()
-        if vim.fn.has('linux') then
-            return nil
-        end
-        if vim.fn.has('win32') then
-            return { "C:/Program Files/Microsoft Visual Studio/2022/Enterprise/VC/Tools/Llvm/bin/clangd.exe" }
-        end
-        return nil
-    end)()
-}
+  }
 
-require('lspconfig').rust_analyzer.setup{
+  require('lspconfig').rust_analyzer.setup{
     on_attach = on_attach,
     flags = lsp_flags(),
     -- Server-specific settings...
-}
+  }
 
-require('lspconfig').lua_ls.setup {
+  require('lspconfig').lua_ls.setup {
     on_attach = on_attach,
     flags = lsp_flags(),
-}
+    settings = {
+      Lua = {
+        diagnostics = {
+          globals = { "vim" }
+        }
+      }
+    }
+  }
 --vim.lsp.start({
   -- name = 'my-server-name',
   -- cmd = {'name-of-language-server-executable'},
   -- root_dir = vim.fs.dirname(vim.fs.find({'setup.py', 'pyproject.toml'}, { upward = true })[1]),
 --})
+end
+
+return {
+  "neovim/nvim-lspconfig",
+  init = function()
+  end,
+  config = lsp_config
+}
