+++ lua/plugins/whichkey.lua
@@ -1,0 +1,32 @@
+return {
+  "folke/which-key.nvim",
+  event = "VeryLazy",
+  init = function()
+    vim.o.timeout = true
+    vim.o.timeoutlen = 300
+  end,
+  config = function()
+    require "which-key".setup {
+      marks = true,
+      registers = true,
+      presets = {
+        operators = true,
+        motions = true,
+        text_objects = true,
+        windows = true,
+        nav = true,
+        z = true,
+        g = true
+      },
+      motions = {
+        count = true
+      },
+      icons = {
+        breadcrumb = "»",
+        separator = "➜",
+        group = "+"
+      },
+
+    }
+  end
+}
